Move UnmodifiableObservableList to commons.core
