+++ src/main/java/seedu/address/logic/commands/SelectCommand.java
@@ -3,7 +3,7 @@ package seedu.address.logic.commands;
 import seedu.address.commons.core.EventsCenter;
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.events.controller.JumpToListRequestEvent;
-import seedu.address.model.UnmodifiableObservableList;
+import seedu.address.commons.core.UnmodifiableObservableList;
 import seedu.address.model.person.ReadOnlyPerson;
 
 /**
