+++ src/main/java/seedu/address/model/ModelManager.java
@@ -2,6 +2,7 @@ package seedu.address.model;
 
 import javafx.collections.transformation.FilteredList;
 import seedu.address.commons.core.LogsCenter;
+import seedu.address.commons.core.UnmodifiableObservableList;
 import seedu.address.commons.util.StringUtil;
 import seedu.address.commons.events.model.ModelChangedEvent;
 import seedu.address.commons.core.ComponentManager;
