+++ src/test/java/seedu/address/model/UnmodifiableObservableListTest.java
@@ -5,6 +5,7 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
+import seedu.address.commons.core.UnmodifiableObservableList;
 
 import java.util.*;
